// src/app/medicamento/models/interfaces/medicamento.interface.ts
// Generated by https://quicktype.io
 
export interface IMedicamento {
    id: number;
    nombre: string;
    descripcion: string;
    precio: number;
    stock: number;
    descuento: number;
    caducidad: string;
    imagen?: string;
    farmacia: IFarmacia;
}
 
export interface IFarmacia {
    id: number;
    nombre: string;
    rfc: string;
    descripcion: string;
    telefono: string;
    direccion?: IDireccion;
}
 
export interface IDireccion {
    id: number;
    calle: string;
    numeroInterior?: number;
    numeroExterior?: number;
    estado: IEstado;
    municipio: IMunicipio;
    colonia: IColonia;
}
 
// Generated by https://quicktype.io
 
export interface IEstado {
    id: number;
    nombre: string;
    descripcion?: string;
}
 
// Generated by https://quicktype.io
 
export interface IMunicipio {
    id: number;
    nombre: string;
    descripcion?: string;
}
 
 
export interface IColonia {
    id: number;
    nombre: string;
    descripcion?: string;
    codigoPostal: string;
}